useExampleList hooks implemented
